change get_eip

+++ dync_hook/func_template.cpp
@@ -16,7 +16,7 @@ extern "C" FUNC_RET_TYPE FUNC_NAME(ARGS_WITH_NAME){
     Space::enterSys();
     void *eip = options::dync_geteip ? get_eip() : 0;
     Space::exitSys();
-    ret = tern_FUNC_NAME((unsigned)(uint64_t) eip, ARGS_ONLY_NAME);
+    ret = tern_FUNC_NAME((uintptr_t) eip, ARGS_ONLY_NAME);
 #else
     ret = tern_FUNC_NAME(ARGS_ONLY_NAME);
 #endif
